feat(analyze/url): thread company context fields into the URL prompt
Frontend 'HELP US UNDERSTAND YOUR COMPANY BETTER' section collects five
optional (one required) context fields that were being dropped. Plumb them
through so Claude can calibrate the 7-dimension scoring against them.

- UrlAnalyzeRequest gains company_description, customers, stage,
  business_model, gtm_goals (all default to "" so unset fields stay empty)
- build_url_user_message appends an 'Additional company context' block
  listing only the non-empty (whitespace-aware) fields; if every extra is
  blank the message reverts to the prior minimal shape
- analyze_url route forwards the five fields into the builder
- Relax MarketSizingBand.pct to float — Claude legitimately returns
  fractional SOM percentages (e.g. 1.5%) that int-strict rejected

+++ src/schemas/gtm.py
@@ -29,7 +29,7 @@ class MarketOpportunity(BaseModel):
 class MarketSizingBand(BaseModel):
     label: str = ""
     value: str = ""
-    pct: int = 0
+    pct: float = 0
     note: str = ""
